feat: add stubs for noncgo systems to increase compatibility
Allow to compile and use go-sysinfo for noncgo systems.
If a feature needs cgo, an error is returned without compromising
the noncgo features.
